requirements.txt
discord.py>=2.3.0
PyNaCl>=1.5.0
davey>=0.1.0
ytmusicapi>=1.11.0
yt-dlp>=2024.04.09
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0
